refactor(brush): move jitter, speed, and texture controls to brush modal
Toolbar now only shows Size, Opacity, Hardness, Fade, and symmetry
toggles. Jitter sliders, Speed Size, and texture controls live in the
brush settings modal under new Dynamics and Texture sections.
