+++ e2e/brush-jitter-texture.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect, type Page } from './fixtures';
-import { setToolOption, setBrushModalOption, closeBrushModal } from './helpers';
+import { setToolOption, setBrushModalOption, openBrushModal, closeBrushModal } from './helpers';
 
 // ---------------------------------------------------------------------------
 // Helpers
@@ -130,21 +130,27 @@ async function setupBrush(page: Page, opts: { size: number; opacity: number; har
 }
 
 async function setJitter(page: Page, sizeJ: number, angleJ: number, opacityJ: number) {
-  await setToolOption(page, 'Size Jitter', sizeJ);
-  await setToolOption(page, 'Angle Jitter', angleJ);
-  await setToolOption(page, 'Opacity Jitter', opacityJ);
+  await openBrushModal(page);
+  await setBrushModalOption(page, 'Size Jitter', sizeJ);
+  await setBrushModalOption(page, 'Angle Jitter', angleJ);
+  await setBrushModalOption(page, 'Opacity Jitter', opacityJ);
+  await closeBrushModal(page);
 }
 
 async function selectTexture(page: Page, textureName: string) {
-  const select = page.locator('role=toolbar >> select[title="Brush texture"]');
+  await openBrushModal(page);
+  const select = page.locator('[role="dialog"][aria-label="Brushes"] select[title="Brush texture"]');
   await select.selectOption({ label: textureName });
   await page.waitForTimeout(100);
+  await closeBrushModal(page);
 }
 
 async function selectTextureBlendMode(page: Page, mode: string) {
-  const select = page.locator('role=toolbar >> select[title="Texture blend mode"]');
+  await openBrushModal(page);
+  const select = page.locator('[role="dialog"][aria-label="Brushes"] select[title="Texture blend mode"]');
   await select.selectOption({ label: mode });
   await page.waitForTimeout(100);
+  await closeBrushModal(page);
 }
 
 // ---------------------------------------------------------------------------
@@ -381,7 +387,8 @@ test.describe('Brush speed size (#346)', () => {
 
     await setupBrush(page, { size: 40, opacity: 100, hardness: 100 });
     await setJitter(page, 0, 0, 0);
-    await setToolOption(page, 'Speed Size', 80);
+    await setBrushModalOption(page, 'Speed Size', 80);
+    await closeBrushModal(page);
 
     // Slow stroke (many steps = slow mouse movement)
     await drawStroke(page, { x: 50, y: 60 }, { x: 350, y: 60 }, 60);
@@ -422,7 +429,8 @@ test.describe('Brush speed size (#346)', () => {
 
     await setupBrush(page, { size: 40, opacity: 100, hardness: 100 });
     await setJitter(page, 0, 0, 0);
-    await setToolOption(page, 'Speed Size', 0);
+    await setBrushModalOption(page, 'Speed Size', 0);
+    await closeBrushModal(page);
 
     // Fast stroke — should still be full width since speed size is disabled
     await drawStroke(page, { x: 50, y: 100 }, { x: 350, y: 100 }, 3);
@@ -549,8 +557,11 @@ test.describe('Brush texture (#346)', () => {
     await setJitter(page, 0, 0, 0);
 
     // Draw with Noise texture at 50% scale (smaller tiles, more oscillation)
-    await selectTexture(page, 'Noise');
-    await setToolOption(page, 'Scale', 50);
+    await openBrushModal(page);
+    const texSelect = page.locator('[role="dialog"][aria-label="Brushes"] select[title="Brush texture"]');
+    await texSelect.selectOption({ label: 'Noise' });
+    await setBrushModalOption(page, 'Scale', 50);
+    await closeBrushModal(page);
     await drawStroke(page, { x: 50, y: 100 }, { x: 350, y: 100 }, 40);
 
     const smallScaleSamples = await readCompositedStrip(page, 100, 80, 320, 2);
@@ -561,7 +572,8 @@ test.describe('Brush texture (#346)', () => {
     await page.keyboard.press('Control+z');
     await page.waitForTimeout(300);
 
-    await setToolOption(page, 'Scale', 200);
+    await setBrushModalOption(page, 'Scale', 200);
+    await closeBrushModal(page);
     await drawStroke(page, { x: 50, y: 100 }, { x: 350, y: 100 }, 40);
 
     const largeScaleSamples = await readCompositedStrip(page, 100, 80, 320, 2);
