+++ src/app/OptionsBar/tool-options/BrushOptions.module.css
@@ -18,20 +18,6 @@
   border-color: var(--color-border-strong);
 }
 
-.separator {
-  width: 1px;
-  height: 20px;
-  background: var(--color-border);
-  flex-shrink: 0;
-  margin: 0 var(--spacing-xs);
-}
-
-.textureGroup {
-  display: flex;
-  gap: var(--spacing-xs);
-  align-items: center;
-}
-
 .symmetryGroup {
   display: flex;
   gap: var(--spacing-xs);
